Added Scope types to manage scoped storage.
Referred to map scope as module scope, because that better reflects the semantics. Each Module gets a separate ModuleScope for each MapScope. This means Modules can stay loaded between maps as they are not altered during execution.
Kept a MapScope type to manage the above and in case there is use in the future for properly map-scoped storage.
Also referred to world scope as hub scope, because that seemed a more properly technical term.