Combined MapScope::scriptStart arguments into a ScriptStartInfo class.
Doing this allows for ScriptStartInfo to implement constructor overloads to allow different scriptStart arguments without needing to overload every scriptStart function. It will also allow for named parameter methods should that be useful in the future.
Also added a ScriptStartFunc parameter, which is a function to execute on the started Thread, if one is started.