Rewrote MapScope script* functions to behave more like Hexen.
They now return a bool as in Hexen. Additionally, actions on existing threads now follows the same rules as Hexen, with respect to which states allow which actions.
There is one known incompatibility. Deferred actions always "succeed". Whereas in Hexen they would fail if an action was already deferred for the given script. ZDoom allows multiple actions to be queued up, and that seems like a more generally useful behavior. However, a way to enable strict behavioral compatibility will be needed eventually.