ACSVM: Rewrite serialization API
New API is no longer tied to std::iostream, and should be easier to
integrate with a wider variety of engines.

I am also retiring the C API rather than update its serialization. It
is Current Year and if a project needs a good ACS interpreter, it can
use C++.